api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Ujwal Jain", email = "reachujwal@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "openai>=1.61.1",
    "python-dotenv>=1.0.1",
    "supabase>=2.13.0",
]

[project.scripts]
api = "api:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
venvPath = "."
venv = ".venv"
